// src/interfaces/catalogos.model.ts
export class catalogosModel {
    Categoria?: string = "";
    ClaveCategoria?: string = "";
    CodigoTabla?: string = "";
    Descripcion?: string = "";
    Monto?: string = "";
    DescripcionCorta?: string = "";
    CodigoAdicional?: string = "";
    CodigoSistema?: string = "";
    Estado?: string = "";
    StrEstado?: string = "";
    Mensaje?: string = "";
    Success?: boolean = false;
    IdUsuario?: number = 0;
    IdUsuarioEncrypt?: string = "";
    NombreUsuario?: string = "";
  }